C# split plan module into AST elements
